로그인 멤버만 좋아요 가능

+++ src/main/java/com/example/demo/service/LikeService.java
@@ -20,10 +20,12 @@ public class LikeService {
 	}
 
 	public void insertLike(int loginedMemberId, int id) {
+		if(loginedMemberId <= 0) return;
 		likeRepository.insertLike(loginedMemberId, id);
 	}
 
 	public void deleteLike(int loginedMemberId, int id) {
+		if(loginedMemberId <= 0) return;
 		likeRepository.deleteLike(loginedMemberId, id);
 		
 	}
